@@ -1,55 +1,60 @@
 {
-    "name": "publiux/laravelcdn",
+    "name": "abdullahfaqeir/laravelcdn",
     "description": "Content Delivery Network (CDN) Package for Laravel",
     "license": "MIT",
     "keywords": [
-      "laravel",
-      "CDN",
-      "upload",
-      "AWS",
-      "S3",
-      "AWS S3",
-      "amazon",
-      "Assets Upload",
-      "Content Delivery Network"
+        "laravel",
+        "CDN",
+        "upload",
+        "AWS",
+        "S3",
+        "AWS S3",
+        "amazon",
+        "Assets Upload",
+        "Content Delivery Network",
+        "devloops"
     ],
     "authors": [
         {
             "name": "Raul Ruiz",
-            "email": "publiux@gmail.com"
+            "email": "abdullahfaqeir@gmail.com"
+        },
+        {
+            "name": "Abdullah Al-Faqeir",
+            "email": "abdullah@devloops.net"
         }
     ],
     "require": {
-      "php": ">=7.2.5",
-      "illuminate/support": "^8|^9",
-      "illuminate/config": "^8|^9",
-      "symfony/finder": "^5",
-      "symfony/console": "^5",
-      "aws/aws-sdk-php": "~3.2",
-      "laravel/framework": "^8|^9"
+        "php": ">=8.2",
+        "illuminate/support": "^8|^9",
+        "illuminate/config": "^8|^9",
+        "symfony/finder": "^5|^6",
+        "symfony/console": "^5|^6",
+        "aws/aws-sdk-php": "~3.2",
+        "laravel/framework": "^8|^9"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.0",
-        "mockery/mockery": "1.4.2"
+        "mockery/mockery": "^1.5.0"
     },
     "autoload": {
         "classmap": [
             "tests/TestCase.php",
-            "src/Publiux/laravelcdn/Exceptions"
+            "src/AbdullahFaqeir/LaravelCDN/Exceptions"
         ],
-      "psr-0": {
-            "Publiux\\laravelcdn\\": "src/"
+        "psr-0": {
+            "AbdullahFaqeir\\LaravelCDN\\": "src/"
         }
     },
-  "extra": {
-    "laravel": {
-      "providers": [
-        "Publiux\\laravelcdn\\CdnServiceProvider"
-      ],
-      "aliases": {
-        "CDN": "Publiux\\laravelcdn\\Facades\\CdnFacadeAccessor"
-      }
-    }
-  },
+    "extra": {
+        "laravel": {
+            "providers": [
+                "AbdullahFaqeir\\LaravelCDN\\CdnServiceProvider"
+            ],
+            "aliases": {
+                "CDN": "AbdullahFaqeir\\LaravelCDN\\Facades\\CdnFacadeAccessor"
+            }
+        }
+    },
     "minimum-stability": "stable"
 }
