@@ -1,26 +1,25 @@
 <?php
 
-
-namespace Publiux\laravelcdn\Tests;
+namespace AbdullahFaqeir\LaravelCDN\Tests;
 
 // requiring this file to reference assertions as global functions. (to skip the $this->)
 require_once 'vendor/phpunit/phpunit/src/Framework/Assert/Functions.php';
 
-use PHPUnit_Framework_TestCase as PHPUnit;
+use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 
-class TestCase extends PHPUnit
+class TestCase extends PHPUnitTestCase
 {
     public function __construct()
     {
         parent::__construct();
     }
 
-    public function setUp()
+    public function setUp(): void
     {
         parent::setUp();
     }
 
-    public function tearDown()
+    public function tearDown(): void
     {
         parent::tearDown();
     }
@@ -28,13 +27,14 @@ class TestCase extends PHPUnit
     /**
      * Call protected/private method of a class.
      *
-     * @param object &$object    Instantiated object that we will run method on.
-     * @param string $methodName Method name to call
-     * @param array  $parameters Array of parameters to pass into method.
+     * @param object &$object     Instantiated object that we will run method on.
+     * @param string  $methodName Method name to call
+     * @param array   $parameters Array of parameters to pass into method.
      *
      * @return mixed Method return.
+     * @throws \ReflectionException
      */
-    public function invokeMethod(&$object, $methodName, array $parameters = [])
+    public function invokeMethod(&$object, string $methodName, array $parameters = []): mixed
     {
         $reflection = new \ReflectionClass(get_class($object));
         $method = $reflection->getMethod($methodName);
